+++ lib/solargraph/rspec/correctors/subject_method_corrector.rb
@@ -16,6 +16,7 @@ module Solargraph
 
             subject_pin = rspec_subject_method(namespace_pin, subject_name, location_range, fake_method_ast)
             add_pin(subject_pin)
+            add_pins(one_liner_expectation_pins(subject_pin))
           end
 
           rspec_walker.after_walk do
@@ -23,7 +24,11 @@ module Solargraph
 
             namespace_pin = closest_namespace_pin(namespace_pins, described_class_pin.location.range.start.line)
 
-            add_pin(implicit_subject_pin(described_class_pin, namespace_pin)) if namespace_pin
+            if namespace_pin
+              implicit_subject_pin = implicit_subject_method(described_class_pin, namespace_pin)
+              add_pin(implicit_subject_pin)
+              add_pins(one_liner_expectation_pins(implicit_subject_pin))
+            end
           end
         end
 
@@ -47,7 +52,7 @@ module Solargraph
         # @param described_class_pin [Pin::Method]
         # @param namespace_pin [Pin::Namespace]
         # @return [Pin::Method]
-        def implicit_subject_pin(described_class_pin, namespace_pin)
+        def implicit_subject_method(described_class_pin, namespace_pin)
           described_class = described_class_pin.return_type.first.subtypes.first.name
 
           PinFactory.build_public_method(
@@ -55,6 +60,41 @@ module Solargraph
             'subject',
             types: ["::#{described_class}"],
             location: described_class_pin.location,
+            scope: :instance
+          )
+        end
+
+        # @param subject_pin [Pin::Method]
+        # @return [Array<Pin::Method>]
+        def one_liner_expectation_pins(subject_pin)
+          [
+            one_liner_expecation_pin(subject_pin.closure, :is_expected, subject_pin.location),
+            one_liner_expecation_pin(subject_pin.closure, :should, subject_pin.location),
+            one_liner_expecation_pin(subject_pin.closure, :should_not, subject_pin.location)
+          ]
+        end
+
+        # @param namespace_pin [Pin::Namespace]
+        # @param method_name [:is_expected, :should, :should_not]
+        # @param location [Solargraph::Location]
+        # @return [Pin::Method]
+        def one_liner_expecation_pin(namespace_pin, method_name, location)
+          return_type = case method_name
+                        when :is_expected
+                          ['::RSpec::Expectations::ExpectationTarget']
+                        when :should
+                          ['::RSpec::Matchers::BuiltIn::PositiveOperatorMatcher']
+                        when :should_not
+                          ['::RSpec::Matchers::BuiltIn::NegativeOperatorMatcher']
+                        else
+                          raise ArgumentError, "Unknown inline expectation method: #{method_name}"
+                        end
+
+          PinFactory.build_public_method(
+            namespace_pin,
+            method_name.to_s,
+            types: [return_type],
+            location: location,
             scope: :instance
           )
         end